refine codes
