+++ research/gaia-x/ir/runtime/src/lib.rs
@@ -13,11 +13,11 @@
 //! See the License for the specific language governing permissions and
 //! limitations under the License.
 
-pub use crate::graph::graph::{get_graph, register_graph};
+pub use crate::graph::{get_graph, register_graph};
 pub use compiler::IRJobCompiler;
 
 mod compiler;
-mod error;
+pub mod error;
 pub mod expr;
 pub mod graph;
 mod process;
